# swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Health Options AI",
        "version": "1.0.0",
        "title": "Health Options AI",
        "contact": {
            "email": "abc@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "schemes": ["http"],
    "host": "localhost:3000",
    "basePath": "",
    "paths" : {
...
        "/" :{
            "post": {
                "operationId": "verifyOtp",
                "summary": "verifyOtp > Request an OTP for login",
                "tags": [ "Auth"],
                "description": "User can request an OTP using either their email or mobile number.",
                "produces": [ "application/json"],
                "consumes": ["application/json"],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "task object",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "routeName" : {
                                    "type": "string",
                                    "default" : "sendOtp"
                                },
                                "email": {
                                    "type": "string",
                                    "format": "email",
                                    "example": "user@example.com"
                                },
                                "mobileNumber": {
                                    "type": "string",
                                    "example": "1234567890"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OTP sent successfully",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "OTP sent successfully"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string",
                                    "example": "Email or mobile number is required"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string",
                                    "example": "Internal server error"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }, 
...
}